requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-autoslug==1.9.8
django-crispy-forms==1.14.0
Pillow==9.4.0
sqlparse==0.4.3
django-extensions